Reorganize into src/services and src/ui
