+++ packages/shared-ux/error_boundary/mocks/src/storybook.ts
@@ -8,8 +8,8 @@
 
 import { AbstractStorybookMock } from '@kbn/shared-ux-storybook-mock';
 import { action } from '@storybook/addon-actions';
-import { ErrorService } from '../../src/error_service';
-import { ToastsService } from '../../src/toasts_service';
+import { ErrorService } from '../../src/services/error_service';
+import { ToastsService } from '../../src/services/toasts_service';
 import { ErrorBoundaryServices } from '../../types';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
