+++ packages/shared-ux/error_boundary/src/services/error_service.ts
